feat: support creating a workspace
