+++ frontend/appflowy_flutter/lib/shared/feature_flags.dart
@@ -17,7 +17,7 @@ enum FeatureFlag {
   bool get isOn {
     switch (this) {
       case FeatureFlag.collaborativeWorkspace:
-        return false;
+        return true;
       case FeatureFlag.membersSettings:
         return false;
     }
