+++ frontend/appflowy_flutter/lib/workspace/presentation/home/menu/sidebar/rename_view_dialog.dart
@@ -15,7 +15,7 @@ import 'package:flutter/material.dart';
 Future<void> createViewAndShowRenameDialogIfNeeded(
   BuildContext context,
   String dialogTitle,
-  void Function(String viewName) createView,
+  void Function(String viewName, BuildContext context) createView,
 ) async {
   final value = await getIt<KeyValueStorage>().getWithFormat(
     KVKeys.showRenameDialogWhenCreatingNewFile,
@@ -27,9 +27,9 @@ Future<void> createViewAndShowRenameDialogIfNeeded(
       title: dialogTitle,
       value: LocaleKeys.menuAppHeader_defaultNewPageName.tr(),
       autoSelectAllText: true,
-      confirm: createView,
+      onConfirm: createView,
     ).show(context);
-  } else {
-    createView(LocaleKeys.menuAppHeader_defaultNewPageName.tr());
+  } else if (context.mounted) {
+    createView(LocaleKeys.menuAppHeader_defaultNewPageName.tr(), context);
   }
 }
