+++ frontend/appflowy_flutter/lib/workspace/presentation/home/menu/sidebar/sidebar_new_page_button.dart
@@ -23,7 +23,7 @@ class SidebarNewPageButton extends StatelessWidget {
       onPressed: () async => createViewAndShowRenameDialogIfNeeded(
         context,
         LocaleKeys.newPageText.tr(),
-        (viewName) {
+        (viewName, _) {
           if (viewName.isNotEmpty) {
             context.read<MenuBloc>().add(MenuEvent.createApp(viewName));
           }
