+++ frontend/appflowy_flutter/lib/workspace/presentation/home/menu/sidebar/sidebar_workspace.dart
@@ -1,17 +1,13 @@
 import 'package:appflowy/generated/flowy_svgs.g.dart';
-import 'package:appflowy/generated/locale_keys.g.dart';
-import 'package:appflowy/shared/af_role_pb_extension.dart';
-import 'package:appflowy/util/color_generator/color_generator.dart';
 import 'package:appflowy/workspace/application/user/user_workspace_bloc.dart';
-import 'package:appflowy/workspace/presentation/home/menu/sidebar/_sidebar_workspace_actions.dart';
 import 'package:appflowy/workspace/presentation/home/menu/sidebar/sidebar_setting.dart';
+import 'package:appflowy/workspace/presentation/home/menu/sidebar/workspace/_sidebar_workspace_icon.dart';
+import 'package:appflowy/workspace/presentation/home/menu/sidebar/workspace/_sidebar_workspace_item_list.dart';
 import 'package:appflowy/workspace/presentation/notifications/widgets/notification_button.dart';
-import 'package:appflowy/workspace/presentation/settings/widgets/members/workspace_member_bloc.dart';
 import 'package:appflowy_backend/protobuf/flowy-folder/view.pb.dart';
 import 'package:appflowy_backend/protobuf/flowy-user/user_profile.pb.dart';
 import 'package:appflowy_editor/appflowy_editor.dart';
 import 'package:appflowy_popover/appflowy_popover.dart';
-import 'package:easy_localization/easy_localization.dart';
 import 'package:flowy_infra_ui/flowy_infra_ui.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
@@ -90,7 +86,7 @@ class _WorkspaceWrapperState extends State<_WorkspaceWrapper> {
                 if (currentWorkspace == null || workspaces.isEmpty) {
                   return const SizedBox.shrink();
                 }
-                return _WorkspaceMenu(
+                return WorkspacesMenu(
                   userProfile: widget.userProfile,
                   currentWorkspace: currentWorkspace,
                   workspaces: workspaces,
@@ -107,7 +103,7 @@ class _WorkspaceWrapperState extends State<_WorkspaceWrapper> {
               const HSpace(4.0),
               SizedBox(
                 width: 24.0,
-                child: _WorkspaceIcon(workspace: widget.currentWorkspace),
+                child: WorkspaceIcon(workspace: widget.currentWorkspace),
               ),
               const HSpace(8),
               FlowyText.medium(
@@ -123,193 +119,5 @@ class _WorkspaceWrapperState extends State<_WorkspaceWrapper> {
       // TODO: Lucas.Xu. mobile workspace menu
       return const Placeholder();
     }
-  }
-}
-
-class _WorkspaceMenu extends StatelessWidget {
-  const _WorkspaceMenu({
-    required this.userProfile,
-    required this.currentWorkspace,
-    required this.workspaces,
-  });
-
-  final UserProfilePB userProfile;
-  final UserWorkspacePB currentWorkspace;
-  final List<UserWorkspacePB> workspaces;
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      mainAxisSize: MainAxisSize.min,
-      children: [
-        // user email
-        Padding(
-          padding: const EdgeInsets.symmetric(vertical: 8, horizontal: 4),
-          child: Row(
-            children: [
-              FlowyText.medium(
-                _getUserInfo(),
-                fontSize: 12.0,
-                overflow: TextOverflow.ellipsis,
-                color: Theme.of(context).hintColor,
-              ),
-              const Spacer(),
-              FlowyButton(
-                useIntrinsicWidth: true,
-                text: const FlowySvg(FlowySvgs.add_m),
-                onTap: () {
-                  // TODO: mock create workspace
-                  context.read<UserWorkspaceBloc>().add(
-                        const UserWorkspaceEvent.createWorkspace(
-                          'Hello World',
-                          'this is a fake workspace description.',
-                        ),
-                      );
-                },
-              ),
-            ],
-          ),
-        ),
-        for (final workspace in workspaces) ...[
-          _WorkspaceMenuItem(
-            workspace: workspace,
-            userProfile: userProfile,
-            isSelected: workspace.workspaceId == currentWorkspace.workspaceId,
-          ),
-          const VSpace(4.0),
-        ],
-      ],
-    );
-  }
-
-  String _getUserInfo() {
-    if (userProfile.email.isNotEmpty) {
-      return userProfile.email;
-    }
-
-    if (userProfile.name.isNotEmpty) {
-      return userProfile.name;
-    }
-
-    return LocaleKeys.defaultUsername.tr();
-  }
-}
-
-class _WorkspaceMenuItem extends StatelessWidget {
-  const _WorkspaceMenuItem({
-    required this.workspace,
-    required this.userProfile,
-    required this.isSelected,
-  });
-
-  final UserProfilePB userProfile;
-  final UserWorkspacePB workspace;
-  final bool isSelected;
-
-  @override
-  Widget build(BuildContext context) {
-    return BlocProvider(
-      create: (_) => WorkspaceMemberBloc(
-        userProfile: userProfile,
-        workspace: workspace,
-      )
-        ..add(const WorkspaceMemberEvent.initial())
-        ..add(const WorkspaceMemberEvent.getWorkspaceMembers()),
-      child: BlocBuilder<WorkspaceMemberBloc, WorkspaceMemberState>(
-        builder: (context, state) {
-          final members = state.members;
-          // settings right icon inside the flowy button will
-          //  cause the popover dismiss intermediately when click the right icon.
-          // so using the stack to put the right icon on the flowy button.
-          return Stack(
-            alignment: Alignment.center,
-            children: [
-              FlowyButton(
-                onTap: () {
-                  if (!isSelected) {
-                    context.read<UserWorkspaceBloc>().add(
-                          UserWorkspaceEvent.openWorkspace(
-                            workspace.workspaceId,
-                          ),
-                        );
-                  }
-                },
-                margin: const EdgeInsets.symmetric(vertical: 8, horizontal: 12),
-                iconPadding: 10.0,
-                leftIconSize: const Size.square(32),
-                leftIcon: _WorkspaceIcon(
-                  workspace: workspace,
-                ),
-                text: Column(
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  children: [
-                    FlowyText.medium(
-                      workspace.name,
-                      fontSize: 14.0,
-                      overflow: TextOverflow.ellipsis,
-                    ),
-                    if (members.length > 1)
-                      FlowyText(
-                        '${members.length} ${LocaleKeys.settings_appearance_members_members.tr()}',
-                        fontSize: 10.0,
-                        color: Theme.of(context).hintColor,
-                      ),
-                  ],
-                ),
-              ),
-              Positioned(
-                right: 12.0,
-                child: Align(child: _buildRightIcon(context)),
-              ),
-            ],
-          );
-        },
-      ),
-    );
-  }
-
-  Widget _buildRightIcon(BuildContext context) {
-    // only the owner can update or delete workspace.
-    // only show the more action button when the workspace is selected.
-    if (!isSelected ||
-        !context.read<WorkspaceMemberBloc>().state.myRole.isOwner) {
-      return const SizedBox.shrink();
-    }
-
-    return Row(
-      children: [
-        WorkspaceMoreActionList(workspace: workspace),
-        const FlowySvg(
-          FlowySvgs.blue_check_s,
-          blendMode: null,
-        ),
-      ],
-    );
-  }
-}
-
-class _WorkspaceIcon extends StatelessWidget {
-  const _WorkspaceIcon({
-    required this.workspace,
-  });
-
-  final UserWorkspacePB workspace;
-
-  @override
-  Widget build(BuildContext context) {
-    // TODO: support icon later
-    return Container(
-      alignment: Alignment.center,
-      decoration: BoxDecoration(
-        color: ColorGenerator.generateColorFromString(workspace.name),
-        borderRadius: BorderRadius.circular(4),
-      ),
-      child: FlowyText(
-        workspace.name.isEmpty ? '' : workspace.name.substring(0, 1),
-        fontSize: 16,
-        color: Colors.black,
-      ),
-    );
   }
 }
