+++ frontend/appflowy_flutter/lib/workspace/presentation/home/menu/view/view_item.dart
@@ -1,5 +1,3 @@
-import 'package:flutter/material.dart';
-
 import 'package:appflowy/generated/flowy_svgs.g.dart';
 import 'package:appflowy/generated/locale_keys.g.dart';
 import 'package:appflowy/plugins/base/emoji/emoji_text.dart';
@@ -25,6 +23,7 @@ import 'package:easy_localization/easy_localization.dart';
 import 'package:flowy_infra_ui/flowy_infra_ui.dart';
 import 'package:flowy_infra_ui/style_widget/hover.dart';
 import 'package:flowy_infra_ui/widget/flowy_tooltip.dart';
+import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
 typedef ViewItemOnSelected = void Function(ViewPB);
@@ -456,7 +455,7 @@ class _SingleInnerViewItemState extends State<SingleInnerViewItem> {
             createViewAndShowRenameDialogIfNeeded(
               context,
               _convertLayoutToHintText(pluginBuilder.layoutType!),
-              (viewName) {
+              (viewName, _) {
                 if (viewName.isNotEmpty) {
                   viewBloc.add(
                     ViewEvent.createView(
@@ -499,7 +498,7 @@ class _SingleInnerViewItemState extends State<SingleInnerViewItem> {
                 autoSelectAllText: true,
                 value: widget.view.name,
                 maxLength: 256,
-                confirm: (newValue) {
+                onConfirm: (newValue, _) {
                   context.read<ViewBloc>().add(ViewEvent.rename(newValue));
                 },
               ).show(context);
