+++ frontend/resources/translations/en.json
@@ -64,7 +64,9 @@
       "exportLogFiles": "Export log files",
       "reachOut": "Reach out on Discord"
     },
-    "deleteWorkspaceHintText": "Are you sure you want to delete the workspace? This action cannot be undone."
+    "deleteWorkspaceHintText": "Are you sure you want to delete the workspace? This action cannot be undone.",
+    "createSuccess": "Workspace created successfully",
+    "createFailed": "Failed to create workspace"
   },
   "shareAction": {
     "buttonText": "Share",
